+++ packages/index/src/index.js
@@ -13,6 +13,8 @@ const { statIndex, prettyPrint } = require('./stat')
 const { matcherFns } = require('./merge');
 const { readStream, readStreams } = require('./read-stream');
 const createLimitFilter = require('./limit-filter');
+const { getJournalFilename, createJournal, readJournal } = require('./journal')
+const { getIndexName } = require('./utils')
 
 const asyncReadIndex = promisify(readIndex)
 const asyncFileFilter = promisify(fileFilter)
@@ -27,11 +29,11 @@ const asyncCreateOrUpdate = async (directory, filename, options) => {
   const fileIndex = await asyncReadIndex(filename)
   const limitFilter = await asyncCreateLimitFilter(fileIndex.data.length, options.addLimits, options.filter)
   const fsEntries = await asyncCreateIndex(directory, {...options, filter: limitFilter})
-  const [entries, changed ] = await asyncUpdateIndex(fileIndex.data, fsEntries, options.matcherFn)
+  const [entries, changes] = await asyncUpdateIndex(fileIndex.data, fsEntries, options.matcherFn)
   const limitExceeded = limitFilter.limitExceeded()
 
-  if (!changed) {
-    return [fileIndex, limitExceeded]
+  if (!changes) {
+    return [fileIndex, limitExceeded, changes]
   }
 
   const newIndex = {
@@ -41,27 +43,26 @@ const asyncCreateOrUpdate = async (directory, filename, options) => {
     data: entries
   }
   if (options.dryRun) {
-    return [newIndex, limitExceeded]
+    return [newIndex, limitExceeded, changes]
   }
   const writeIndex = await asyncWriteIndex(filename, newIndex);
-  return [writeIndex, limitExceeded]
+  return [writeIndex, limitExceeded, changes]
 }
 
-const asyncUpdateChecksum = async (filename, index, updateChecksums, isDryRun) => {
-  if (!updateChecksums) {
-    return index
+const asyncUpdateChecksum = async (filename, index, withChecksum, isDryRun) => {
+  if (!withChecksum) {
+    return [index, false]
   }
 
   const sha1sumDate = new Date().toISOString();
-  const [updateIndex, changed] = await asyncChecksum(index, sha1sumDate)
+  const [updateIndex, checksumChanges] = await asyncChecksum(index, sha1sumDate)
 
-  if (!changed || isDryRun) {
-    return index
+  if (!checksumChanges || isDryRun) {
+    return [index, checksumChanges]
   }
 
   const writeIndex = await asyncWriteIndex(filename, updateIndex);
-  updateIndex = null
-  return writeIndex
+  return [writeIndex, checksumChanges];
 }
 
 const asyncUpdate = async (directory, filename, options) => {
@@ -69,15 +70,20 @@ const asyncUpdate = async (directory, filename, options) => {
   debug(`Updating file index for directory ${directory}`);
 
   const filter = await asyncFileFilter(options.exclude, options.excludeFromFile)
-  const [index, limitExceeded] = await asyncCreateOrUpdate(directory, filename, {...options, filter})
-  const updateIndex = await asyncUpdateChecksum(filename, index, options.checksum, options.dryRun)
+  const [index, limitExceeded, changes] = await asyncCreateOrUpdate(directory, filename, {...options, filter})
+  const [updateIndex, checksumChanges] = await asyncUpdateChecksum(filename, index, options.checksum, options.dryRun)
+  await createJournal(filename, updateIndex, changes, checksumChanges, options.journal, options.dryRun)
+
   debug(`Updated file index for directory ${directory} in ${Date.now() - t0}ms`);
   return [updateIndex, limitExceeded]
 }
 
 module.exports = {
+  getIndexName,
+  getJournalFilename,
   readStream,
   readStreams,
+  readJournal: callbackify(readJournal),
   update: callbackify(asyncUpdate),
   matcherFns,
   prettyPrint,