+++ packages/index/src/utils.js
@@ -1,0 +1,7 @@
+const path = require('path')
+
+const getIndexName = filename => path.basename(filename).replace(/\.[^.]+$/, '')
+
+module.exports = {
+  getIndexName
+}