+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add end to end tests
+- index: Add file index journal via option `--journal`
 
 ## [1.1.0] - 2021-06-22